pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "SNV-FEAST"
version = "0.0.8"
authors = [
  { name="Leah Briscoe, Eran Halperin, Nandita Garud", email="leahpbriscoe@gmail.com" },
]
description = "A package to get signature SNVs from a given strain and set of sink and sources, using MIDAS output"
readme = "README.md"
requires-python = ">=3.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = ['numpy>=1.20.3','scipy>=1.7.1','pyyaml>=6.0','pandas>=1.3.4','bz2file>=0.98']

[project.urls]
"Homepage" = "https://github.com/garudlab/SNV-FEAST"
"Bug Tracker" = "https://github.com/garudlab/SNV-FEAST/issues"
